Fix on this page list generation for config
On config pages there were headings without links in them. The script
errored on that and didn't generate a table of contents.
Fix that error by adding anchors to some headings.

Not all headings will always have link in them though, so when that
happens, ignore the heading so that the script will not error.