Fix navigation active states
Pages with an index.html wouldn't show their active state properly (not
bold). Active pages with sub-pages would also not open their sub-pages
list. In both scenarios it was unclear in the navigation which pages was
open.

Fix this by removing the `index.html` from the links in the
`navigation.yml` definition. That's because the active link helpers
don't expect a `.html` for an index/overview file.