Add guides section with deploy markers guide
Add first guide about deploy markers to the new guides section in the
docs. This section will be more hands on in taking people through the
steps we recommend they take to get the most out of AppSignal.

I started with the deploy markers guide because that would be my first
recommendation after an app reports data to AppSignal.