Add session data filtering guide (#386)
Merge the Ruby and Elixir session data filtering pages into one guide.
I couldn't find any information about this "feature" in the JavaScript
section, and couldn't really tell if it's entirely implemented based
on their code bases, so I omitted this integration for now. Feel free
to add it if you are sure how it works!

I added a new page that explains the session data filtering in some more
detail.

I've moved the "filter out all session data" part of the Ruby and Elixir
pages to the /application/session-data-filtering.html page because it
didn't feel like the best place for it was the guide. The "feature" page
seems like a better fit, for when people are reading further up on it.