Merge ignore errors pages
Merge the pages from the integrations that had separate ignore error
pages.

I didn't add a specific feature page for "errors". We should do that in
the future, but right now that would be a lot of work and we can gain
more by focussing on restructuring the content we already have.

I couldn't find one for the front-end JavaScript integration, so I wrote
it based on the information on the configuration page and the source
code.