Link to relevant guides and sections
The customize data collection page is/was a guide page already. Now that
most of the content are separate guides, link to those instead.

This page can probably be removed at some point when we have a guide or
feature page for tagging. And decide how to mention what we do with
query filtering.