Add ignore namespaces guide (#393)
We already had a namespaces guide, but it didn't go into how to ignore
a namespaces. This guide, in the filter data topic, does explain it and
is linked to from the namespaces guide. It's not part of that guide,
because setting up namespaces and ignoring them are two different use
cases.