Fix markdown helper error on first build (#397)
When a page is loaded that uses the markdown helper, in a view that's
nor markdown, middleman would raise an error. This would occur in an
`.erb` for the integration options pages if it was the first page to be
requested after middleman server started.

This could also be a fix for #269, where pages were built in parallel
and caused a similar error. There too it could be that one of the build
processes rendered such as page first, causing the error.

Fix it by setting the `scope` class attribute on the AppsignalMarkdown
class if it's not present to the current view object.
This way the AppsignalMarkdown class can call view helpers, instead of
erroring on a `nil` scope object.