Document Ruby gem 3.0 changes, add upgrade guide
Update docs to mention changes in the Ruby 3.0 gem.

Add an upgrade guide to help users guide them through the upgrade. The
most important step though is probably upgrading to the latest 2.x
series gem as that prints and logs warnings for all the things that may
break in Ruby gem 3.0.

Closes #418