Fix Node.js config option list case format
For the Node.js integration some config options use the camelCase
format. The docs config options page wasn't originally build to handle
that, so it would render the config option list in camelCase and the
anchor for the config option details on the page in all lowercase.

Update the config options list to also render the link with a lowercase
anchor name so it links to the right part of the page.