Fix anchors on sample guide page
Prefix repeating headings for different language with the language
prefix.

This way the `#params` link generated by the multiple `### params`
headings don't all link to the same/first occurrence on the page.

By prefixing the heading with our custom markdown extension `^` we can
modify the heading anchor generated for the heading.