Rewrite AppSignal naming terminology section
The naming for all packages are similar enough to not say they are
different.

Where in Elixir we have an `appsignal` package, we also have
`appsignal_plug` and `appsignal_phoenix` packages.

In Node.js we only have prefixed packages like `appsignal/nodejs`, but
the naming is still that "appsignal" is the package name for both
Node.js and Elixir. There are just also subpackages under that prefix.