Bump Ruby version to 2.7.5
The Ruby 2.4.x series has been End Of Life since 2020-03-31.
Update it to a more recent version that is still supported.

I'm not updating to Ruby 3.0.x, because that might be a bigger change
and I don't want to do that now. I know it works on 2.7.x, because
I've been running it locally for a while now.