Move Node.js minutely probes to instrumentation
The Node.js has a "metrics" section that other integrations do not have.
Minutely probes are grouped in the "custom instrumentation" section for
all other integrations, so let's move the Node.js page there too.

Remove the metrics page, that was already merged with the global custom
metrics page, so we can remove the entire section. No content that isn't
documented somewhere else is lost by removing this content.