Document Ember 4 workaround
This PR updates the documentation to Ember 4, documenting the
"workaround" for Ember 4 as the default way to use the integration.
This should not be a problem for users of previous Ember versions
as using it this way was already possible in those versions.

It also updates the information on `ember-auto-import`, as Ember 4
includes it on its `ember new` template by default.